requirements.txt
# Production dependencies
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-cors-headers==4.4.0
django-filter==24.3
django-safedelete==1.4.1
psycopg[binary]==3.2.3
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
Pillow==11.0.0